web/requirements.txt
psycopg==3.1.13
pandas==1.5.1
numpy==1.23.4
python-dotenv==1.0.0
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1


